fix(apkd_dbus_server): fix reporting errors during operations to dbus clients
We previously catched exceptions early in ApkInterfacer since we didn't
have proper error handling in our methodHandler. This has changed
in the meantime, so we should stop catching exceptions in ApkInterfacer
since we'll catch them in methodHandler and send a DBus error to our
clients so they know something went wrong.