feat(apkd): reword error messages
Previously error messages were rather short and didn't
really help the user solve problems with apk much.
With this rework, apk-polkit will tell the user
what command to run in the CLI to diagnose the
problem. Unfortunately it's not really possible to
provide this via a GUI since solving conflicts can
be rather hard (and hopefully they won't happen
to casual users, which are the main users of this).
In theory we *could* run `apk fix` for good measure
and try again in case solving goes wrong (as this
could fix things), but this could also go very south
since `apk fix` also re-evaluates the world, so it
might also purge (vital) packages from underneath
you with no indication of doing so to the user
via the GUI. As such I think it's best for the user
to explicitly run `apk fix` if necessary. Again,
hopefully users won't have to do this.

Also don't use apk_error_str to get the error
messages of solver operations. The integer
returned by the apk_solver* functions we use _does not_
indicate which error occured but only indicates
_the number_ of errors that occured. This could
lead to very weird error messages, e.g.
if you had one error in a transaction, it'd say
"access denied" as error messages when it actually
just wanted to express that there has been 1 error
during the operation.

fixes #15